refactor: integrate frontend with i2Sales API
