+++ services/interactions.ts
@@ -1,4 +1,4 @@
-import { apiClient } from './apiClient';
+import { apiClient } from './api';
 import { Interaction } from '../types';
 
 export async function createInteraction(
